# integration-kit/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "reports-fetcher",
    "version": "1.0.1",
    "description": "Corrected, self-contained HTTP contract for reports-fetcher v1.0.1 (baseline 76fee13). This file fixes the gaps in the auto-generated schema: the mandatory Idempotency-Key request header, the 200 idempotent-replay response, RFC 9457 problem+json error responses, and the binary file response. It is the authoritative machine-readable contract for the local mock and for production. Mock-only routes (/__mock/*) and the X-Mock-Scenario header are intentionally NOT part of this contract.",
    "contact": {
      "name": "reports-fetcher integration kit"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Production loopback (or an SSH tunnel / gateway mapped to it)"
    },
    {
      "url": "http://127.0.0.1:18765",
      "description": "Local mock kit"
    }
  ],
...
  "paths": {
...
    "/api/v1/reports": {
      "get": {
        "tags": ["reports"],
        "summary": "List archived reports (read-only, never fetches)",
        "operationId": "list_reports",
        "parameters": [
          {"name": "market", "in": "query", "required": false,
           "schema": {"type": "string", "enum": ["CN", "HK", "US"]}},
          {"name": "symbol", "in": "query", "required": false,
           "schema": {"type": "string"}},
          {"name": "doc_type", "in": "query", "required": false,
           "schema": {"type": "string"}},
          {"name": "period_from", "in": "query", "required": false,
           "schema": {"type": "string", "format": "date"}},
          {"name": "period_to", "in": "query", "required": false,
           "schema": {"type": "string", "format": "date"}},
          {"name": "limit", "in": "query", "required": false,
           "schema": {"type": "integer", "minimum": 1, "maximum": 100,
                      "default": 20}},
          {"name": "cursor", "in": "query", "required": false,
           "description": "Opaque cursor from a previous next_cursor",
           "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Page of archived reports",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ReportListOut"}
              }
            }
          },
          "400": {
            "description": "Invalid cursor",
            "content": {
              "application/problem+json": {
                "schema": {"$ref": "#/components/schemas/Problem"}
              }
            }
          },
          "422": {
            "description": "Invalid filter value",
            "content": {
              "application/problem+json": {
                "schema": {"$ref": "#/components/schemas/Problem"}
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "tags": ["reports"],
        "summary": "Report metadata and file versions",
        "operationId": "get_report",
        "parameters": [
          {"name": "report_id", "in": "path", "required": true,
           "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Report detail",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ReportDetailOut"}
              }
            }
          },
          "404": {
            "description": "Report not found",
            "content": {
              "application/problem+json": {
                "schema": {"$ref": "#/components/schemas/Problem"}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Application token. Local loopback mode runs without authentication."
      }
    },
    "schemas": {
...
      "ReportListItem": {
        "type": "object",
        "properties": {
          "report_id": {"type": "string"},
          "market": {"type": "string", "enum": ["CN", "HK", "US"]},
          "symbol": {"type": "string"},
          "issuer_id": {"type": ["string", "null"]},
          "source_id": {"type": "string"},
          "source_url": {"type": ["string", "null"]},
          "title": {"type": ["string", "null"]},
          "doc_type": {"type": "string"},
          "report_period": {"type": ["string", "null"], "format": "date"},
          "period_source": {"type": "string",
                            "enum": ["source_field", "explicit_title",
                                     "unknown"]},
          "filing_date": {"type": ["string", "null"], "format": "date"},
          "language": {"type": ["string", "null"]},
          "is_amendment": {"type": "boolean", "default": false},
          "status": {"type": "string"},
          "artifact_id": {"type": ["string", "null"]},
          "sha256": {"type": ["string", "null"]},
          "bytes": {"type": ["integer", "null"]},
          "fetched_at": {"type": ["string", "null"], "format": "date-time"},
          "warnings": {"type": "array", "items": {"type": "string"},
                       "default": []},
          "download_url": {"type": "string"}
        },
        "required": ["report_id", "market", "symbol", "source_id", "doc_type",
                     "period_source", "status", "download_url"]
      },
      "ReportListOut": {
        "type": "object",
        "properties": {
          "items": {"type": "array",
                    "items": {"$ref": "#/components/schemas/ReportListItem"}},
          "next_cursor": {"type": ["string", "null"]}
        },
        "required": ["items"]
      },
      "ArtifactOut": {
        "type": "object",
        "properties": {
          "artifact_id": {"type": "string"},
          "sha256": {"type": "string"},
          "bytes": {"type": "integer"},
          "media_type": {"type": "string",
                         "enum": ["application/pdf", "text/html"]},
          "fetched_at": {"type": "string", "format": "date-time"},
          "state": {"type": "string",
                    "enum": ["staged", "ready", "unavailable"]},
          "is_current": {"type": "boolean"}
        },
        "required": ["artifact_id", "sha256", "bytes", "media_type",
                     "fetched_at", "state", "is_current"]
      },
      "ReportDetailOut": {
        "type": "object",
        "properties": {
          "report_id": {"type": "string"},
          "market": {"type": "string", "enum": ["CN", "HK", "US"]},
          "symbol": {"type": "string"},
          "source_id": {"type": "string"},
          "source_url": {"type": ["string", "null"]},
          "title": {"type": ["string", "null"]},
          "doc_type": {"type": "string"},
          "source_form": {"type": ["string", "null"]},
          "report_period": {"type": ["string", "null"], "format": "date"},
          "period_source": {"type": "string",
                            "enum": ["source_field", "explicit_title",
                                     "unknown"]},
          "filing_date": {"type": ["string", "null"], "format": "date"},
          "language": {"type": ["string", "null"]},
          "is_amendment": {"type": "boolean", "default": false},
          "revision_of": {"type": ["string", "null"]},
          "status": {"type": "string"},
          "current_artifact_id": {"type": ["string", "null"]},
          "warnings": {"type": "array", "items": {"type": "string"},
                       "default": []},
          "artifacts": {"type": "array",
                        "items": {"$ref": "#/components/schemas/ArtifactOut"}}
        },
        "required": ["report_id", "market", "symbol", "source_id", "doc_type",
                     "period_source", "status", "artifacts"]
      },
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details extended with stable code, request_id, retryable and optional field errors.",
        "properties": {
          "type": {"type": "string", "default": "about:blank"},
          "title": {"type": "string"},
          "status": {"type": "integer"},
          "detail": {"type": "string"},
          "instance": {"type": "string"},
          "code": {"type": "string"},
          "request_id": {"type": "string"},
          "retryable": {"type": "boolean"},
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {"type": "string"},
                "detail": {"type": "string"}
              }
            }
          }
        },
        "required": ["title", "status", "code", "request_id", "retryable"]
      }
    }
  }
}